+++ packages/mocha/tests/it.always-test.js
@@ -11,23 +11,27 @@ describe('BigTest Mocha: it.always', () => {
     tests = results.tests;
   }));
 
-  it('successfully passes before the timeout', () => {
-    expect(tests[0].duration).to.be.within(80, 100);
+  it('successfully passes after the timeout', () => {
+    expect(tests[0].duration).to.be.within(100, 120);
     expect(tests[0].err).to.be.empty;
   });
 
-  it('throws when the assertion fails', () => {
-    expect(tests[1].duration).to.be.within(100, 120);
-    expect(tests[1].err).to.have.property('expected', '0');
-  });
-
   it('can modify the timeout', () => {
-    expect(tests[2].duration).to.be.within(30, 50);
-    expect(tests[2].err).to.be.empty;
+    expect(tests[1].duration).to.be.within(50, 70);
+    expect(tests[1].err).to.be.empty;
   });
 
-  it('.only has multiple aliases', () => {
-    expect(convergentIt.always.only).to.equal(convergentIt.only.always);
+  it('throws when the assertion fails', () => {
+    expect(tests[2].duration).to.be.within(200, 220);
+    expect(tests[2].err).to.have.property('expected', '0');
+  });
+
+  it('always sets the test timeout to 0', () => {
+    let test = convergentIt.always('test', () => {});
+
+    return test.timeout(50).fn().then(() => {
+      expect(test.timeout()).to.equal(0);
+    });
   });
 });
 
@@ -38,9 +42,13 @@ describe('BigTest Mocha: it.always.only', () => {
     tests = results.tests;
   }));
 
+  it('has multiple aliases', () => {
+    expect(convergentIt.always.only).to.equal(convergentIt.only.always);
+  });
+
   it('runs a single test', () => {
     expect(tests).to.have.lengthOf(1);
-    expect(tests[0].duration).to.be.within(80, 100);
+    expect(tests[0].duration).to.be.within(100, 120);
     expect(tests[0].err).to.be.empty;
   });
 });