+++ packages/mocha/tests/it.pause-test.js
@@ -1,4 +1,4 @@
-import { describe, before, it } from 'mocha';
+import { describe, beforeEach, it } from 'mocha';
 import { expect } from 'chai';
 
 import { it as convergentIt } from '../src';
@@ -6,16 +6,17 @@ import { it as convergentIt } from '../src';
 describe('BigTest Mocha: it.pause', () => {
   let test;
 
-  before(() => {
+  beforeEach(() => {
     test = convergentIt.pause('test');
   });
 
-  it('should set the timeout to 0', () => {
-    expect(test._timeout).to.equal(0);
+  it('returns a promise', () => {
+    expect(test.fn()).to.be.a('Promise');
   });
 
-  it('should return a promise', () => {
-    expect(test.fn()).to.be.a('Promise');
+  it('always sets the test timeout to 0', () => {
+    test.timeout(2000).fn();
+    expect(test.timeout()).to.equal(0);
   });
 
   it('.only has multiple aliases', () => {